Rename domain dir core to avoid confusion with gis terminology
