+++ nad_ch/core/entities.py
